requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
opencv-python
streamlit
pillow
python-dotenv
gdown